# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LibreFang API",
    "description": "LibreFang Agent Operating System — REST API for managing AI agents, tools, workflows, and more.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "2026.4.28-beta7"
  },
  "paths": {
...
    "/a2a/tasks/send": {
      "post": {
        "tags": [
          "a2a"
        ],
        "summary": "POST /a2a/tasks/send — Submit a task to an agent via A2A.",
        "operationId": "a2a_send_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Submit a task to an agent via A2A",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/a2a/tasks/{id}": {
      "get": {
        "tags": [
          "a2a"
        ],
        "summary": "GET /a2a/tasks/{id} — Get task status from the task store.",
        "operationId": "a2a_get_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get A2A task status",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/a2a/tasks/{id}/cancel": {
      "post": {
        "tags": [
          "a2a"
        ],
        "summary": "POST /a2a/tasks/{id}/cancel — Cancel a tracked task.",
        "operationId": "a2a_cancel_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancel a tracked A2A task",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/a2a/agents/{id}": {
      "get": {
        "tags": [
          "a2a"
        ],
        "summary": "GET /api/a2a/agents/{id} — Get a specific external A2A agent by index, URL, or name.",
        "operationId": "a2a_get_external_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a specific external A2A agent",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/a2a/discover": {
      "post": {
        "tags": [
          "a2a"
        ],
        "summary": "POST /api/a2a/discover — Discover a new external A2A agent by URL.",
        "operationId": "a2a_discover_external",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discover an external A2A agent by URL",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/a2a/send": {
      "post": {
        "tags": [
          "a2a"
        ],
        "summary": "POST /api/a2a/send — Send a task to an external A2A agent.",
        "operationId": "a2a_send_external",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Send a task to an external A2A agent",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/a2a/tasks/{id}/status": {
      "get": {
        "tags": [
          "a2a"
        ],
        "summary": "GET /api/a2a/tasks/{id}/status — Get task status from an external A2A agent.",
        "operationId": "a2a_external_task_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "url",
            "in": "query",
            "description": "URL of the external A2A agent",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get external A2A task status",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "GET /api/agents — List agents with optional filtering, pagination, and sorting.",
        "description": "Query parameters (all optional — omitting them returns all agents):\n  - `q`: free-text search across name and description (case-insensitive)\n  - `status`: filter by lifecycle state (e.g. \"running\", \"suspended\")\n  - `limit` / `offset`: pagination\n  - `sort`: field to sort by — \"name\", \"created_at\", \"last_active\", \"state\"\n  - `order`: \"asc\" (default) or \"desc\"",
        "operationId": "list_agents",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Free-text search on name/description",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by agent state",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max items to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Items to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field: name, created_at, last_active, state",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order: asc or desc",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of agents"
          }
        }
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "POST /api/agents — Spawn a new agent.",
        "operationId": "spawn_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpawnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent spawned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpawnResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid manifest"
          }
        }
      }
    },
    "/api/agents/bulk": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "POST /api/agents/bulk — Create multiple agents at once.",
        "operationId": "bulk_create_agents",
        "requestBody": {
          "description": "Array of agent spawn requests",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Create multiple agents at once",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "DELETE /api/agents/bulk — Delete multiple agents at once.",
        "operationId": "bulk_delete_agents",
        "requestBody": {
          "description": "Array of agent IDs to delete",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAgentIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Delete multiple agents at once",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/agents/bulk/start": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "POST /api/agents/bulk/start — Set multiple agents to Full mode.",
        "operationId": "bulk_start_agents",
        "requestBody": {
          "description": "Array of agent IDs to start",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAgentIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Start multiple agents (set to Full mode)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/agents/bulk/stop": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "POST /api/agents/bulk/stop — Stop multiple agents' current runs.",
        "operationId": "bulk_stop_agents",
        "requestBody": {
          "description": "Array of agent IDs to stop",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAgentIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stop multiple agents' current runs",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/agents/{id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "GET /api/agents/:id — Get a single agent's detailed info.",
        "operationId": "get_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent details",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Agent not found"
          }
        }
      },
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "DELETE /api/agents/:id — Kill an agent.",
        "operationId": "kill_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent killed"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      },
      "patch": {
        "tags": [
          "agents"
        ],
        "operationId": "patch_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Partial agent fields to update",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Partially update an agent (name, description, model, system prompt)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/agents/{id}/clone": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "POST /api/agents/{id}/clone — Clone an agent with its workspace files.",
        "operationId": "clone_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "New name for the cloned agent",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Clone an agent with its workspace files",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/agents/{id}/config": {
      "patch": {
        "tags": [
          "agents"
        ],
        "summary": "PATCH /api/agents/{id}/config — Hot-update agent name, description, system prompt, and identity.",
        "operationId": "patch_agent_config",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Agent config fields to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAgentConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Hot-update agent name, description, system prompt, identity, and model",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/agents/{id}/files": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "GET /api/agents/{id}/files — List workspace identity files.",
        "operationId": "list_agent_files",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List workspace identity files for an agent",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/agents/{id}/history": {
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "DELETE /api/agents/{id}/history — Clear ALL conversation history for an agent.",
        "operationId": "clear_agent_history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clear all conversation history for an agent",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/agents/{id}/identity": {
      "patch": {
        "tags": [
          "agents"
        ],
        "summary": "PATCH /api/agents/{id}/identity — Update an agent's visual identity.",
        "operationId": "update_agent_identity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Identity fields to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update an agent's visual identity",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SpawnRequest": {
        "type": "object",
        "description": "Request to spawn an agent from a TOML manifest string or a template name.",
        "properties": {
          "manifest_toml": {
            "type": "string",
            "description": "Agent manifest as TOML string (optional if `template` is provided)."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional custom name for the agent. Overrides the name from the template\nor manifest, allowing multiple agents from the same template."
          },
          "signed_manifest": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional Ed25519 signed manifest envelope (JSON).\nWhen present, the signature is verified before spawning."
          },
          "template": {
            "type": [
              "string",
              "null"
            ],
            "description": "Template name from `~/.librefang/workspaces/agents/{template}/agent.toml`.\nWhen provided and `manifest_toml` is empty, the template is loaded automatically."
          }
        }
      },
      "SpawnResponse": {
        "type": "object",
        "description": "Response after spawning an agent.",
        "required": [
          "agent_id",
          "name"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}